// frontend/src/views/affect/AF/AWAF00603S/type.d.ts
/** ----------------------------------------------------------------
 * 業務名称　: 養鶏-互助防疫システム
 * 機能概要　: 共通バー情報
 * 　　　　　  インターフェース定義
 * 作成日　　: 2023.04.17
 * 作成者　　: 屠
 * 変更履歴　:
 * -----------------------------------------------------------------*/
import { Enum共通バー番号 } from '#/Enums'

//-------------------------------------------------------------------
//リクエスト
//-------------------------------------------------------------------
/** 初期化処理 */
export interface InitRequest extends DaRequestBase {
  /** 機能ID */
  kinoid: string
  /** 宛名番号 */
  atenano: string
  /** パターンNo.(ドロップダウンリストコード) */
  patternno?: string
}

//-------------------------------------------------------------------
//レスポンス
//-------------------------------------------------------------------
/** 初期化処理 */
export interface InitResponse extends DaResponseBase {
  /** 共通バー情報 */
  kekkalist: InfoVM[]
}

//-------------------------------------------------------------------
//ビューモデル
//-------------------------------------------------------------------
/** 共通バー情報 */
export interface InfoVM {
  /** 共通バー番号 */
  barno: Enum共通バー番号
  /** 共通バー名称 */
  barnm: string
  /** 注意フラグ */
  attnflg: boolean
  /** アクセス権限(権限がない場合:false) */
  enabled: boolean
  /** 追加権限フラグ */
  addflg: boolean
  /** 修正権限フラグ */
  updateflg: boolean
  /** 削除権限フラグ */
  deleteflg: boolean
  /** 個人番号利用権限フラグ */
  personalnoflg: boolean
}